First pass at exposing management capabilities to client
A set of boolean values that exposes to the client what actions can be taken for management.
This moves the logic from the client side to the cloud side.

This is just a first pass as the code for exposing these booleans and enforcing them cloud side needs to be unified.