Removed "auth-" prefix from everywhere
This is a major overhaul of the file structure of the project.  All of the modules were originally prefixed with "auth-" but were also living directly in the "com.unitvectory' package.  After moving them into the "com.unitvectory.serviceauthcentral" package the longer module names didn't make much sense anymore.

This is an improvement but obviously this is a big change impacting almost every file in the repo.