Renaming the Java package
The original package name was "com.unitvectory.auth" which didn't match the maven structure.  This renames that to "com.unitvectory.serviceauthcentral" which matches the package name for Maven much better.

Again, this is a huge number of files changing across the project, but no real changes other than renaming the package which requires moving the files in the folder structure.