Including "description" claim in JWTs for users
When a user is logged in it would be very helpful to know which user that is as identified by their description.
While this isn't perfect right now as the description is populated with a not-perfect value.  This is a major improvement to the UI as it is now clear to a user which account they are logged in with and by placing it in the JWT this doesn't require an additional API call to retrieve this information.