Fixed test cases and enhanced to allow for testing reads
The GraphQL test cases hadn't been fleshed out, this makes it so that an appropriate mock JWT can be passed in based on the access token and also adds an additional test for getting a client that was created.

The test cases here still need to be fleshed out more.  The memory database isn't reset for each test which means they can step on one another.  Also the overall coverage here is really poor.